Cleanup password interface
