+++ app/core/interfaces/password/input.html
@@ -1,0 +1,14 @@
+<input type="password" value="{{value}}" name="{{name}}" class="medium password-primary" style="display:block;margin-bottom:10px;" placeholder="{{t "password_placeholder"}}" spellcheck="false" autocomplete="off" autocorrect="off" autocapitalize="off"/>
+<input type="password" id="password_fake" class="hidden" autocomplete="off" style="display: none;">
+
+<span class="password-text"></span>
+
+{{#if require_confirmation}}
+<input type="password" value="{{value}}" class="medium password-confirm" style="display:block;margin-bottom:10px;" placeholder="{{t "password_confirm_placeholder"}}" spellcheck="false" autocomplete="off" autocorrect="off" autocapitalize="off"/>
+{{/if}}
+
+<div style="display:block;">
+	<button class="button password-generate" style="margin-right:10px;" type="button">{{t "generate_new"}}</button>
+	<button class="button password-toggle" style="display:none;" type="button">{{t "reveal_password"}}</button>
+	<span class="encrypted hidden">{{t "password_encrypted"}}</span>
+</div>
